+++ modules/pathkit/tests/pathops.spec.js
@@ -1,4 +1,4 @@
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 20000;
 
 var dumpErrors = false;
 var container;
@@ -84,7 +84,7 @@ describe('PathKit\'s PathOps Behavior', function() {
         } else {
             PathKitInit({
                 locateFile: (file) => '/pathkit/'+file,
-            }).then((_PathKit) => {
+            }).ready().then((_PathKit) => {
                 PathKit = _PathKit;
                 PATHOP_MAP = {
                     'kIntersect_SkPathOp':         PathKit.PathOp.INTERSECT,