+++ src/core/SkOverdrawCanvas.cpp
@@ -61,40 +61,36 @@ SkOverdrawCanvas::SkOverdrawCanvas(SkCanvas* canvas)
     fPaint.setColorFilter(SkColorFilter::MakeMatrixFilterRowMajor255(kIncrementAlpha));
 }
 
-void SkOverdrawCanvas::drawPosTextCommon(const void* text, size_t byteLength, const SkScalar pos[],
+void SkOverdrawCanvas::drawPosTextCommon(const SkGlyphID glyphs[], int count, const SkScalar pos[],
                                          int scalarsPerPos, const SkPoint& offset,
-                                         const SkPaint& paint) {
+                                         const SkFont& font, const SkPaint& paint) {
     ProcessOneGlyphBounds processBounds(this);
     SkSurfaceProps props(0, kUnknown_SkPixelGeometry);
     this->getProps(&props);
     auto cache = SkStrikeCache::FindOrCreateStrikeExclusive(
-                                SkFont::LEGACY_ExtractFromPaint(paint), paint, props,
+                                font, paint, props,
                                 SkScalerContextFlags::kNone, this->getTotalMatrix());
-    SkFindAndPlaceGlyph::ProcessPosText(paint.getTextEncoding(),
-                                        (const char*)text, byteLength,
+    SkFindAndPlaceGlyph::ProcessPosText(glyphs, count,
                                         SkPoint::Make(0, 0), SkMatrix(), (const SkScalar*) pos, 2,
                                         cache.get(), processBounds);
 }
 
 void SkOverdrawCanvas::onDrawTextBlob(const SkTextBlob* blob, SkScalar x, SkScalar y,
                                       const SkPaint& paint) {
-    SkPaint runPaint = paint;
     SkTextBlobRunIterator it(blob);
     for (;!it.done(); it.next()) {
-        size_t textLen = it.glyphCount() * sizeof(uint16_t);
         const SkPoint& offset = it.offset();
-        it.applyFontToPaint(&runPaint);
         switch (it.positioning()) {
             case SkTextBlobRunIterator::kDefault_Positioning:
                 SK_ABORT("This canvas does not support draw text.");
                 break;
             case SkTextBlobRunIterator::kHorizontal_Positioning:
-                this->drawPosTextCommon(it.glyphs(), textLen, it.pos(), 1,
-                                        SkPoint::Make(x, y + offset.y()), runPaint);
+                this->drawPosTextCommon(it.glyphs(), it.glyphCount(), it.pos(), 1,
+                                        SkPoint::Make(x, y + offset.y()), it.font(), paint);
                 break;
             case SkTextBlobRunIterator::kFull_Positioning:
-                this->drawPosTextCommon(it.glyphs(), textLen, it.pos(), 2, SkPoint::Make(x, y),
-                                        runPaint);
+                this->drawPosTextCommon(it.glyphs(), it.glyphCount(), it.pos(), 2, {x, y},
+                                        it.font(), paint);
                 break;
             case SkTextBlobRunIterator::kRSXform_Positioning:
                 // unimplemented ...