+++ src/core/SkPicturePlayback.cpp
@@ -78,7 +78,7 @@ public:
                     fCount = 0;
                 } else {
                     fCount = SkPaintPriv::ValidCountText(fText, fByteLength,
-                                                         paint->getTextEncoding());
+                                                         SkPaintPriv::GetEncoding(*paint));
                     reader->validate(fCount > 0);
                 }
             }
@@ -461,7 +461,7 @@ void SkPicturePlayback::handleOp(SkReadBuffer* reader,
             if (paint && text.text()) {
                 SkFont font = SkFont::LEGACY_ExtractFromPaint(*paint);
                 auto blob = SkTextBlob::MakeFromPosText(text.text(), text.length(), pos, font,
-                                                        paint->getTextEncoding());
+                                                        SkPaintPriv::GetEncoding(*paint));
                 canvas->drawTextBlob(blob, 0, 0, *paint);
             }
         } break;
@@ -479,7 +479,7 @@ void SkPicturePlayback::handleOp(SkReadBuffer* reader,
             if (top < clip.fBottom && bottom > clip.fTop && paint && text.text()) {
                 SkFont font = SkFont::LEGACY_ExtractFromPaint(*paint);
                 auto blob = SkTextBlob::MakeFromPosText(text.text(), text.length(), pos, font,
-                                                        paint->getTextEncoding());
+                                                        SkPaintPriv::GetEncoding(*paint));
                 canvas->drawTextBlob(blob, 0, 0, *paint);
             }
         } break;
@@ -495,7 +495,7 @@ void SkPicturePlayback::handleOp(SkReadBuffer* reader,
             if (paint && text.text()) {
                 SkFont font = SkFont::LEGACY_ExtractFromPaint(*paint);
                 auto blob = SkTextBlob::MakeFromPosTextH(text.text(), text.length(), xpos, constY,
-                                                         font, paint->getTextEncoding());
+                                                         font, SkPaintPriv::GetEncoding(*paint));
                 canvas->drawTextBlob(blob, 0, 0, *paint);
             }
         } break;
@@ -515,7 +515,7 @@ void SkPicturePlayback::handleOp(SkReadBuffer* reader,
             if (top < clip.fBottom && bottom > clip.fTop && paint && text.text()) {
                 SkFont font = SkFont::LEGACY_ExtractFromPaint(*paint);
                 auto blob = SkTextBlob::MakeFromPosTextH(text.text(), text.length(), xpos, constY,
-                                                         font, paint->getTextEncoding());
+                                                         font, SkPaintPriv::GetEncoding(*paint));
                 canvas->drawTextBlob(blob, 0, 0, *paint);
             }
         } break;
@@ -624,15 +624,18 @@ void SkPicturePlayback::handleOp(SkReadBuffer* reader,
             uint32_t flags = reader->readUInt();
             TextContainer text(reader, paint);
             const SkRSXform* xform = (const SkRSXform*)reader->skip(count, sizeof(SkRSXform));
-            const SkRect* cull = nullptr;
             if (flags & DRAW_TEXT_RSXFORM_HAS_CULL) {
-                cull = (const SkRect*)reader->skip(sizeof(SkRect));
+                // skip past cull rect
+                (void)reader->skip(sizeof(SkRect));
             }
             reader->validate(count == text.count());
             BREAK_ON_READ_ERROR(reader);
 
             if (text.text()) {
-                canvas->drawTextRSXform(text.text(), text.length(), xform, cull, *paint);
+                SkFont font = SkFont::LEGACY_ExtractFromPaint(*paint);
+                auto blob = SkTextBlob::MakeFromRSXform(text.text(), text.length(), xform, font,
+                                                        SkPaintPriv::GetEncoding(*paint));
+                canvas->drawTextBlob(blob, 0, 0, *paint);
             }
         } break;
         case DRAW_VERTICES_OBJECT: {