+++ experimental/canvaskit/CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.3.1] - 2019-01-04
+### Added
+ - `SkFont` now exposed.
+ - `MakeCanvasSurface` can now take a canvas element directly.
+ - `MakeWebGLCanvasSurface` can now take a WebGL context as an integer and use it directly.
+
+### Changed
+ - `CanvasKitInit(...).then()` is no longer the recommended way to initialize things.
+It will be removed in 0.4.0. Use `CanvasKitInit(...).ready()`, which returns a real Promise.
+
+### Removed
+- `SkPaint.measureText` - use `SkFont.measureText` instead.
+
 ## [0.3.0] - 2018-12-18
 
 ### Added